rl_ml/pyproject.toml
[project]
name = "rl-forex-trader"
version = "0.1.0"
description = "RL-based Forex Trading Agent using SAC for EUR/USD trading"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Forex RL Trader", email = "trader@example.com"},
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "ta>=0.10.2",
    "stable-baselines3[extra]>=2.0.0",
    "gymnasium>=0.28.0",
    "torch>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "scikit-learn>=1.3.0",
    "pytz>=2023.3",
    "tqdm>=4.66.0",
    "tensorboard>=2.13.0",
]

[project.scripts]
rl-train = "rl_ml.train:main"
rl-test = "rl_ml.test:main"
rl-data = "rl_ml.data_fetcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rl_ml"]